refactor: consolidate onboarding component props definitions and remove unused types
