+++ src/features/onboarding/components/onboarding-mission-card.tsx
@@ -4,15 +4,22 @@ import MissionAchievementStatus from "@/components/mission/mission-achievement-s
 import { Button } from "@/components/ui/button";
 import { Card, CardFooter, CardHeader, CardTitle } from "@/components/ui/card";
 import { MissionIcon } from "@/components/ui/mission-icon";
-import type { OnboardingMissionCardProps } from "@/features/onboarding/types/types";
 import { calculateMissionXp } from "@/features/user-level/utils/level-calculator";
+import type { Tables } from "@/lib/types/supabase";
 import clsx from "clsx";
 import { motion } from "framer-motion";
 import { UsersRound } from "lucide-react";
 
+// オンボーディングミッションカードのProps
+export interface OnboardingMissionCardProps {
+  mission: Omit<Tables<"missions">, "slug">;
+  achievementsCount?: number;
+  userAchievementCount?: number;
+  onCardClick?: () => void;
+}
+
 export default function OnboardingMissionCard({
   mission,
-  achieved,
   achievementsCount,
   userAchievementCount = 0,
   onCardClick,
