+++ src/features/onboarding/components/onboarding-mission-details.tsx
@@ -2,9 +2,15 @@ import { Button } from "@/components/ui/button";
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
 import { DifficultyBadge } from "@/components/ui/difficulty-badge";
 import { MissionIcon } from "@/components/ui/mission-icon";
-import type { OnboardingMissionDetailsProps } from "@/features/onboarding/types/types";
+import type { MockMission } from "@/features/onboarding/types/types";
 import { BUTTON_TEXT, STYLE_CLASSES } from "../constants/constants";
 import { sanitizeHtml } from "../utils/utils";
+
+export interface OnboardingMissionDetailsProps {
+  mission: MockMission;
+  isSubmissionCompleted: boolean;
+  onSubmit: () => void;
+}
 
 /**
  * オンボーディングミッション詳細コンポーネント
