+++ src/features/onboarding/types/types.ts
@@ -1,6 +1,3 @@
-import type { Tables } from "@/lib/types/supabase";
-import type { ReactNode } from "react";
-
 /**
  * オンボーディング機能で使用する型定義（期日前投票専用）
  */
@@ -21,43 +18,6 @@ export interface MockMission {
   is_featured: boolean;
   is_hidden: boolean;
   ogp_image_url: string | null;
-}
-
-// オンボーディングモーダルのProps
-export interface OnboardingModalProps {
-  open: boolean;
-  onOpenChange: (open: boolean) => void;
-}
-
-// オンボーディングボタンのProps
-export interface OnboardingButtonProps {
-  children?: ReactNode;
-  variant?: "default" | "outline" | "link";
-  className?: string;
-}
-
-// オンボーディングキャラクターのProps
-export interface OnboardingCharacterProps {
-  currentDialogue: number;
-  isAnimating: boolean;
-  onNext: () => void;
-  onScrollDown: () => void;
-}
-
-// オンボーディングミッションカードのProps
-export interface OnboardingMissionCardProps {
-  mission: Omit<Tables<"missions">, "slug">;
-  achieved: boolean;
-  achievementsCount?: number;
-  userAchievementCount?: number;
-  onCardClick?: () => void;
-}
-
-// オンボーディングミッション詳細のProps
-export interface OnboardingMissionDetailsProps {
-  mission: MockMission;
-  isSubmissionCompleted: boolean;
-  onSubmit: () => void;
 }
 
 // フック型
